Move MySQL binary API tests to pytest
