+++ .github/workflows/mindsdb.yml
@@ -85,7 +85,11 @@ jobs:
 
           # MySQL API
           echo -e "\n===============test MySQL API===============\n"
-          pytest -vx tests/integration_tests/flows/test_mysql_api_pytest_based.py
+          pytest -vx tests/integration_tests/flows/test_mysql_api.py
+          #
+          # MySQL binary API
+          echo -e "\n===============test MySQL binary API===============\n"
+          pytest -vx -k 'not TestMySqlApi' tests/integration_tests/flows/test_mysql_bin_api.py
 
           mkdir -p ~/.ssh/
           echo "$DB_MACHINE_KEY" > ~/.ssh/db_machine
@@ -96,10 +100,6 @@ jobs:
           sudo chmod 644 ~/.mindsdb_credentials.json
 
           export USE_EXTERNAL_DB_SERVER="1"
-
-          # MySQL binary API
-          echo -e "\n===============\ntest MySQL binary API\n===============\n"
-          python tests/integration_tests/flows/test_mysql_bin_api.py
 
           # Company independent
           echo -e "\n===============\ntest company independent\n===============\n"
