+++ tests/integration_tests/flows/test_mysql_api.py
@@ -1,28 +1,29 @@
+import os
 import time
 import tempfile
-import unittest
 import json
 from pathlib import Path
 
 import docker
-import netifaces
 import pandas as pd
 import requests
+import pytest
+
+from .conftest import docker_inet_ip, TEMP_DIR
 
 
-from common import (
-    run_environment,
-    EXTERNAL_DB_CREDENTIALS,
-    CONFIG_PATH)
+OVERRIDE_CONFIG = {
+    'integrations': {},
+    'api': {
+        "http": {"host": docker_inet_ip()},
+        "mysql": {"host": docker_inet_ip()}
+    }
+}
+# used by (required for) mindsdb_app fixture in conftest
+API_LIST = ["http", "mysql"]
 
-
-def get_docker0_inet_ip():
-    if "docker0" not in netifaces.interfaces():
-        raise Exception("Unable to find 'docker' interface. Please install docker first.")
-    return netifaces.ifaddresses('docker0')[netifaces.AF_INET][0]['addr']
-
-
-HTTP_API_ROOT = f'http://{get_docker0_inet_ip()}:47334/api'
+# used by (required for) mindsdb_app fixture in conftest
+HTTP_API_ROOT = f'http://{docker_inet_ip()}:47334/api'
 
 
 class Dlist(list):
@@ -41,15 +42,56 @@ class Dlist(list):
         return None
 
 
-class TestScenario:
+class BaseStuff:
+    """Contais some helpful set of methods and attributes for tests execution."""
     predictor_name = 'home_rentals'
     file_datasource_name = "from_files"
 
-    def create_datasource(self, db_type):
+    @staticmethod
+    def to_dicts(response):
+        if not response:
+            return {}
+        lines = response.splitlines()
+        if len(lines) < 2:
+            return {}
+        headers = tuple(lines[0].split("\t"))
+        res = Dlist()
+        for body in lines[1:]:
+            data = tuple(body.split("\t"))
+            res.append(dict(zip(headers, data)))
+        return res
+
+    def query(self, _query, encoding='utf-8'):
+        print("EXECUTING BASESTUFF QUERY METHOD!!!!")
+        """Run mysql docker container
+           Perform connection to mindsdb database
+           Execute sql request
+           ----------------------
+           It is very problematic (or even impossible)
+           to provide sql statement as it is in 'docker run command',
+           that's why this action is splitted on three steps:
+               Save sql statement into temporary dir in .sql file
+               Run docker container with volume points to this temp dir,
+               Provide .sql file as input parameter for 'mysql' command"""
+        with tempfile.TemporaryDirectory() as tmpdirname:
+            with open(f"{tmpdirname}/test.sql", 'w') as f:
+                f.write(_query)
+            cmd = f"{self.launch_query_tmpl} < /temp/test.sql"
+            cmd = 'sh -c "' + cmd + '"'
+            res = self.docker_client.containers.run(
+                self.mysql_image,
+                command=cmd,
+                remove=True,
+                volumes={str(tmpdirname): {'bind': '/temp', 'mode': 'ro'}},
+                environment={"MYSQL_PWD": self.config["api"]["mysql"]["password"]})
+        return self.to_dicts(res.decode(encoding))
+
+    def create_datasource(self, db_data):
+        db_type = db_data["type"]
         _query = "CREATE DATASOURCE %s WITH ENGINE = '%s', PARAMETERS = %s;" % (
             db_type.upper(),
             db_type,
-            json.dumps(self.db_creds[db_type]))
+            json.dumps(db_data["connection_data"]))
 
         self.query(_query)
         # and try to drop one of the datasources
@@ -82,8 +124,7 @@ class TestScenario:
             if 'Tables_in_files' in res and res.get_record('Tables_in_files', ds_name):
                 break
             time.sleep(0.5)
-        self.assertTrue('Tables_in_files' in res and res.get_record('Tables_in_files', ds_name),
-                        f"file datasource {ds_name} is not ready to use after {timeout} seconds")
+        assert 'Tables_in_files' in res and res.get_record('Tables_in_files', ds_name), f"file datasource {ds_name} is not ready to use after {timeout} seconds"
 
     def check_predictor_readiness(self, predictor_name):
         timeout = 600
@@ -98,69 +139,93 @@ class TestScenario:
                 elif res.get_record('status', 'error'):
                     raise Exception(res[0]['error'])
             time.sleep(2)
-        self.assertTrue('status' in res and res.get_record('status', 'complete'),
-                        f"predictor {predictor_name} is not complete after {timeout} seconds")
+        assert 'status' in res and res.get_record('status', 'complete'), f"predictor {predictor_name} is not complete after {timeout} seconds"
 
-    def validate_datasource_creation(self, ds_type):
-        self.create_datasource(ds_type.lower())
+    def validate_datasource_creation(self, db_data):
+        ds_type = db_data["type"]
         res = self.query("SELECT * FROM mindsdb.datasources WHERE name='{}';".format(ds_type.upper()))
-        self.assertTrue("name" in res and res.get_record("name", ds_type.upper()),
-                        f"Expected datasource is not found after creation - {ds_type.upper()}: {res}")
+        assert "name" in res and res.get_record("name", ds_type.upper()), f"Expected datasource is not found after creation - {ds_type.upper()}: {res}"
 
-    def test_1_create_datasources(self):
-        for ds_type in ['postgres', 'mysql', 'mariadb']:
-            with self.subTest(msg=ds_type):
-                print(f"\nExecuting {self._testMethodName} ({__name__}.{self.__class__.__name__}) [{ds_type}]")
-                self.validate_datasource_creation(ds_type)
 
-    def test_2_create_predictor(self):
-        _query = f"CREATE PREDICTOR {self.predictor_name} from MYSQL (select * from test_data.home_rentals) as hr_ds predict rental_price;"
+@pytest.mark.usefixtures("mindsdb_app")
+class TestMySqlApi(BaseStuff):
+    """Test mindsdb mysql api.
+    All sql commands are being executed through a docker container with mysql client within.
+    In general all tests do next:
+        1. Do some preconditions
+        2. Specify SQL query needs to be executed
+        3. Launch a docker container with mysql client
+        4. Send the query to the container
+        5. Mysql client inside the container connects to mindsdb mysql api
+        and execute the query"""
+
+    @classmethod
+    def setup_class(cls):
+
+        cls.docker_client = docker.from_env()
+        cls.mysql_image = 'mysql'
+        cls.config = json.loads(Path(os.path.join(TEMP_DIR, "config.json")).read_text())
+
+        cls.launch_query_tmpl = "mysql --host=%s --port=%s --user=%s --database=mindsdb" % (
+            cls.config["api"]["mysql"]["host"],
+            cls.config["api"]["mysql"]["port"],
+            cls.config["api"]["mysql"]["user"])
+
+    @classmethod
+    def tear_down(cls):
+        cls.docker_client.close()
+
+    def test_create_postgres_datasources(self, postgres_db):
+        self.create_datasource(postgres_db)
+        self.validate_datasource_creation(postgres_db)
+
+    def test_create_mariadb_datasources(self, maria_db):
+        self.create_datasource(maria_db)
+        self.validate_datasource_creation(maria_db)
+
+    def test_create_mysql_datasources(self, mysql_db):
+        self.create_datasource(mysql_db)
+        self.validate_datasource_creation(mysql_db)
+
+    def test_create_predictor(self, mysql_db):
+        _query = f"CREATE PREDICTOR {self.predictor_name} from MYSQL (select * from test.rentals) as hr_ds predict rental_price;"
         self.query(_query)
         self.check_predictor_readiness(self.predictor_name)
 
-    def test_3_making_prediction(self):
+    def test_making_prediction(self):
         _query = f"""
             SELECT rental_price, rental_price_explain
             FROM {self.predictor_name}
             WHERE number_of_rooms = 2 and sqft = 400 and location = 'downtown' and days_on_market = 2 and initial_price= 2500;
         """
         res = self.query(_query)
-        self.assertTrue('rental_price' in res and 'rental_price_explain' in res,
-                        f"error getting prediction from {self.predictor_name} - {res}")
+        assert 'rental_price' in res and 'rental_price_explain' in res, f"error getting prediction from {self.predictor_name} - {res}"
 
-    def test_4_describe_predictor_attrs(self):
-        attrs = ["model", "features", "ensemble"]
-        for attr in attrs:
-            with self.subTest(msg=attr):
-                print(f"\nExecuting {self._testMethodName} ({__name__}.{self.__class__.__name__}) [{attr}]")
-                self.query(f"describe mindsdb.{self.predictor_name}.{attr};")
+    @pytest.mark.parametrize("describe_attr",["model", "features", "ensemble"])
+    def test_describe_predictor_attrs(self, describe_attr):
+        self.query(f"describe mindsdb.{self.predictor_name}.{describe_attr};")
 
-    def test_5_service_requests(self):
-        service_requests = [
-            "show databases;",
-            "show schemas;",
-            "show tables;",
-            "show tables from mindsdb;",
-            "show full tables from mindsdb;",
-            "show variables;",
-            "show session status;",
-            "show global variables;",
-            "show engines;",
-            "show warnings;",
-            "show charset;",
-            "show collation;",
-            "show datasources;",
-            "show predictors;",
-            "show function status where db = 'mindsdb';",
-            "show procedure status where db = 'mindsdb';",
-            # "show table status like commands;",
-        ]
-        for req in service_requests:
-            with self.subTest(msg=req):
-                print(f"\nExecuting {self._testMethodName} ({__name__}.{self.__class__.__name__}) [{req}]")
-                self.query(req)
+    @pytest.mark.parametrize("service_req",["show databases;",
+                                        "show schemas;",
+                                        "show tables;",
+                                        "show tables from mindsdb;",
+                                        "show full tables from mindsdb;",
+                                        "show variables;",
+                                        "show session status;",
+                                        "show global variables;",
+                                        "show engines;",
+                                        "show warnings;",
+                                        "show charset;",
+                                        "show collation;",
+                                        "show datasources;",
+                                        "show predictors;",
+                                        "show function status where db = 'mindsdb';",
+                                        "show procedure status where db = 'mindsdb';"]
+                            )
+    def test_service_requests(self, service_req):
+        self.query(service_req)
 
-    def test_7_train_predictor_from_files(self):
+    def test_train_predictor_from_files(self):
         df = pd.DataFrame({
             'x1': [x for x in range(100, 210)] + [x for x in range(100, 210)],
             'x2': [x * 2 for x in range(100, 210)] + [x * 3 for x in range(100, 210)],
@@ -178,11 +243,11 @@ class TestScenario:
         self.query(_query)
         self.check_predictor_readiness(file_predictor_name)
 
-    def test_8_0_select_from_files(self):
+    def test_select_from_files(self):
         _query = f"select * from files.{self.file_datasource_name};"
         self.query(_query)
 
-    def test_9_ts_train_and_predict(self):
+    def test_ts_train_and_predict(self, subtests):
         train_df = pd.DataFrame({
             'group': ["A" for _ in range(100, 210)] + ["B" for _ in range(100, 210)],
             'order': [x for x in range(100, 210)] + [x for x in range(200, 310)],
@@ -224,86 +289,12 @@ class TestScenario:
                 2),
         ]
         for predictor_name, create_query, select_query, res_len in params:
-            with self.subTest(msg=predictor_name):
-                print(f"\nExecuting {self._testMethodName} ({__name__}.{self.__class__.__name__}) [{predictor_name}]")
+            with subtests.test(msg=predictor_name,
+                               predictor_name=predictor_name,
+                               create_query=create_query,
+                               select_query=select_query,
+                               res_len=res_len):
                 self.query(create_query % predictor_name)
                 self.check_predictor_readiness(predictor_name)
                 res = self.query(select_query % predictor_name)
-                self.assertTrue(len(res) == res_len, f"prediction result {res} contains more that {res_len} records")
-
-
-class MySqlApiTest(unittest.TestCase, TestScenario):
-
-    @classmethod
-    def setUpClass(cls):
-        override_config = {
-            'integrations': {},
-            'api': {
-                "http": {"host": get_docker0_inet_ip()},
-                "mysql": {"host": get_docker0_inet_ip()}
-            }
-        }
-
-        run_environment(apis=['http', 'mysql'], override_config=override_config)
-        cls.docker_client = docker.from_env()
-        cls.mysql_image = 'mysql'
-
-        cls.config = json.loads(Path(CONFIG_PATH).read_text())
-
-        with open(EXTERNAL_DB_CREDENTIALS, 'rt') as f:
-            cls.db_creds = json.load(f)
-
-        cls.launch_query_tmpl = "mysql --host=%s --port=%s --user=%s --database=mindsdb" % (
-            cls.config["api"]["mysql"]["host"],
-            cls.config["api"]["mysql"]["port"],
-            cls.config["api"]["mysql"]["user"])
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.docker_client.close()
-
-    def query(self, _query, encoding='utf-8'):
-        """Run mysql docker container
-           Perform connection to mindsdb database
-           Execute sql request
-           ----------------------
-           It is very problematic (or even impossible)
-           to provide sql statement as it is in 'docker run command',
-           that's why this action is splitted on three steps:
-               Save sql statement into temporary dir in .sql file
-               Run docker container with volume points to this temp dir,
-               Provide .sql file as input parameter for 'mysql' command"""
-        with tempfile.TemporaryDirectory() as tmpdirname:
-            with open(f"{tmpdirname}/test.sql", 'w') as f:
-                f.write(_query)
-            cmd = f"{self.launch_query_tmpl} < /temp/test.sql"
-            cmd = 'sh -c "' + cmd + '"'
-            res = self.docker_client.containers.run(
-                self.mysql_image,
-                command=cmd,
-                remove=True,
-                volumes={str(tmpdirname): {'bind': '/temp', 'mode': 'ro'}},
-                environment={"MYSQL_PWD": self.config["api"]["mysql"]["password"]})
-        return self.to_dicts(res.decode(encoding))
-
-    @staticmethod
-    def to_dicts(response):
-        if not response:
-            return {}
-        lines = response.splitlines()
-        if len(lines) < 2:
-            return {}
-        headers = tuple(lines[0].split("\t"))
-        res = Dlist()
-        for body in lines[1:]:
-            data = tuple(body.split("\t"))
-            res.append(dict(zip(headers, data)))
-        return res
-
-
-if __name__ == "__main__":
-    try:
-        unittest.main(failfast=True, verbosity=2)
-        print('Tests passed!')
-    except Exception as e:
-        print(f'Tests Failed!\n{e}')
+                assert len(res) == res_len, f"prediction result {res} contains more that {res_len} records"
